Add language attributes to individual texts
getSomeText() is renamed to getSomeTerm() and returns the full term, not
just the text, and restart() attaches a term in the node and gets text
and xml:lang from there.

We have to store the term because I suspect getSomeTerm() is not
strictly deterministic according to spec – iterating over object members
– and we should re-get it each time because the first time we might not
have had entity data yet.