Add support for reverse links
We now explore the graph in both directions, finding incoming links via
the “what links here” API of MediaWiki. (Unfortunately, that API only
supports one title at a time, so this is somewhat less efficient than
forward links.) getInverseTriples() is mostly the inverse of
getTriples(), except for some peculiarities (e. g. using an API that is
based on titles rather than entity IDs).

Note that we explore links in both direction, but only from the starting
node(s). That is, we don’t explore the outgoing links of nodes found via
incoming links, or vice versa, which means that the graph will generally
form a kind of “hourglass” shape with the starting node(s) in the
center. (Of course, the two halves of the hourglass are not necessarily
balanced in any way, so it might not always look like an hourglass.)