Support more than one predicate
This turns out to be fairly simple to implement – replace the single
predicate with an array of predicates everywhere where it’s just passed
through, and then iterate over it in the two places where the single
predicate was previously actually used.

The URL pattern is changed from …&predicate=… to …&predicates=…, but old
URLs are still accepted.

Implements #7.