Add property ID as title to links
In SVG, instead of the title global attribute there is the title child
element, which may be added to any element to provide a textual
description. We add the property ID of the link, which is especially
useful when the user specified multiple properties (see #7).

Unfortunately, displaying the title requires pixel-perfect mouse
positioning *exactly* on the line (at least in Firefox), but I’m not
sure how to improve that…