Include subentities when searching inverse triples
The backlinks API is page-based, not entity-based, so to include links
from or to sub-entities on the same page, we need to search them
explicitly. Fortunately they’re already included in the entity data, so
we can just raise them to the top-level “entities” response and then
they’re searched automatically.