Support form IDs in entity ID ⇔ title conversion
The backlinks API will happily accept titles with fragments, so for that
it’s okay if we ask for the backlinks of Lexeme:L123#L123-F3, though the
results won’t be any different as if we just omitted the fragment (but
different from what the previous code produced: Lexeme:L123-F3 is not an
existing page!). The other half, support in title2entityId, is not
currently necessary as far as I’m aware, but I like to have those
functions as a symmetric pair that doesn’t lose any information.