Support continuation for backlinks query
By default, Mediawiki only gives us 10 backlinks, and we’re supposed to
make another API request to continue the search, with the parameters
returned in the response. Implementing this correctly is a bit of a
headache, but the new getBacklinks() function should be correct.

Fixes #8 as far as I can tell (graph shows more lexemes now).