Properly propagate promises
Instead of having a bunch of dangling promises with ham-fisted error
handlers, chain them all together by making functions return their
promises and returning those promises in the .then() handlers. This
means we only have to do the error handling once, at the top of the
chain. (Admittedly, the actual handler is still just as ham-fisted, but
at least it can now be improved in one central location.) This also lets
us get rid of callbackDone: the functions are now done when their
promise resolves (with undefined).