Support concurrent getEntities() calls better
Previously, if you made concurrent calls to getEntities(), it was
possible that the entity data would be fetched several times, because we
only cached entity data once we had it. Now, we also store promises in
the cache as soon as we’ve made a request, and concurrent requests await
those promises (in addition to any own requests they’ve made).

This requires that no one else uses the `entities` object directly (only
getEntities() has the necessary logic to handle the promises in there).
As far as I can tell that is the case, except for getSomeTerm(), which
wants to use entity data only if it was already fetched – add a
peekEntity() function for that.

Unfortunately, we can’t move `entities` directly above getEntities(),
because the hoisted getEntities() may be called via start() before an
assignment above getEntities() had a chance to run – so we have to leave
the assignment near the top. But I think it still makes sense to have it
as a separate block from the other variables, at least.