Add legend with subject/predicate labels
This adds a “legend” area below the form, linking to the root entities
and link properties, along with their labels (if available). If there
are several properties, then they’re shown in the same color as used in
the graph itself (I hope all the d3 category10 colors have sufficient
contrast against a white background – I tried setting only the underline
color with text-decoration-color but that was practically invisible).

Fixes #11.