Use property label in arrow title
Previously, we only showed the property ID in the title of each arrow
(link, connection, line). Now that we’re loading the properties anyway
(see commit b63bc671b3), we may as well use their labels instead of the
property IDs.

This also led me to realize that the code for nodes was recreating the
node contents from scratch with each restart() call, which seems
inefficient. I’m pretty sure I cobbled this code together from some d3
examples and tweaked it until it worked, but I think we can do better:
create most of the contents only once (on the “enter” selection [1],
i.e. the new nodes) and only update the texts afterwards, just as we do
for the links.

[1]: https://github.com/d3/d3-selection#selection_enter