Place <title> next to <text> rather than within
Previous DOM structure of a node:

    <g>
      <circle></circle>
      <a>
        <text>
	  <title></title>
	</text>
      </a>
    </g>

Now:

    <g>
      <circle></circle>
      <a>
        <text></text>
	<title></title>
      </a>
    </g>

This doesn’t make a difference for the user as far as I can tell, but it
means that setting the .text() of the <text> no longer removes the
<title>, so now we can create the <title> only once when creating the
node, and all the later DOM changes when updating the nodes after term
data was loaded only affect text nodes, rather than creating or
deleting DOM elements. Hopefully this further improves performance.