Add and show outer terms for senses
This renders senses as “lexeme term (sense term)” instead of only “sense
term”; the outer and proper (inner) term are separate elements with
potentially different language codes (<span> in HTML, <tspan> in SVG).

Note that we don’t fetch the outer entity yet, so unless it happens to
be loaded for some other reason, the outer term will just be the ID with
“und” language code.

Part of #15.