Add pan-and-zoom support
Fortunately this is easy to add following [1], even on the old d3
version we’re on (in newer d3 zoom() is available directly on d3 rather
than on behavior, see [2]). Fixes #16.

Note that the extra <g> in the SVG is very much necessary – adding the
transform directly on the <svg> doesn’t work well.

[1]: https://coderwall.com/p/psogia/simplest-way-to-add-zoom-pan-on-d3-js
[2]: https://d3js.org/d3-zoom