Make svg width 100%
The height needs to be limited a bit to leave space for the form above
the graph, but there should be nothing next to it horizontally, so
there’s no reason for the width to be anything less than 100%. (Note
that, due to browser default styles [e.g. Firefox: margin: 8px on
<body>], this 100% may still be less than 100vw.) For the force layout,
we do need the width in absolute terms, so shuffle the width/height code
around a bit: put the rules directly on the <svg> in the source code and
turn them into absolute integers at initialization time. (This means the
force layout size won’t update if the window is resized later, even
though the <svg> will update, but as far as I can tell this mostly works
out in practice.)

This also removes the old minimum-200 values from the width and height –
I’m not convinced these are necessary anymore: with panning and zooming
available (#16), I think even a graph smaller than 200×200px is not
totally unusable.