Fix reset()
I accidentally broke this in 15a3c52f78 – submitting the form again
would add a new graph but also leave the old one behind, oops. With the
new <g> element in the SVG, I think the cleanest reset is to just drop
the <g> entirely and recreate it; the zoom behavior should probably only
be set once, though. Part of #16.